WIP - rework config model to dedupe context
Rework the configuration model to drop the config into a canary hash in the context to align with what will get
submitted to Mine. Prevents some needless reshuffling in orca, and dedupes the context values when the canary is
in progress.

WIP - this totally busts the rendering of the stage in progress currently