COPY TO CLIPBOARD!!!!!!!!!!!!!!!!!
Copies text to your clipboard without the use of Flash.
Right now it on permalinks on Clusters, Tasks, and Pipelines

AND

In the instance details pane by the instance id you can copy a bash script that will log you into that instance.

NOTE: This doesn't work with the MGMT account yet.