SPIN-247:  Add cluster selector with some smarts.
Both the region list and cluster list are dirived from the what account is selected.
The cluster list starts off with all clusters available.  Once an account is selected the region and cluster list are filter to show only those regions and clusters in the account. Once the region is selected the cluster list will filter down to show only clusters in that region.

Changing either the account or region selections will remove the selecte cluster item if there was one.