SPIN-1353: Make hitting <enter> key will run the execution.
Previously when the <enter> key was hit the close button was 'clicked' because it was the first in the DOM, unless the 'Run' key had focus.

I added a type="button" to the cancel button which Chrome, FF, Safari (only ones tested) will ignore it and use the type="submit" for the <enter> key event.