Adding some right hand room to the multiselect ui-select box
The issue was that the scroll bar in the modal cluster view (clone, create, edit in pipeline) was covering the delete icon for the load balancer and security group multiselect box.

I trimed the width down by 10px and styled the scrollbar in the modals to be thinner and not expand on hover.