SPIN-1921: Fix upsert of projects. (#2508)
Upserting a project doesn't correctly pull the project name out of the taskCommand when creating the URL for the POST method.  This cases a 405 Method Not Allowed error.

Adding the project to the taskCommand payload to the taskExecutor.  Also checking for project name when constructing the url for the POST call.