fix(deck): Add check for isZeroLengthText response
Ran into an issue when saving pipelines where I would submit a request
to save and the response coming back had a Content-Type of text/plain
which would cause the request to "fail" regardless of if it actually
succeeded or not. Adds a check to allow zero length text plain headers
as well.