fix(provider/kubernetes): fixes bug with edit server group dialog delays (#3719)
When the pipeline has a docker registry trigger the value of c.imageDescription.imageId contains the (Tag resolved at runtime) from the trigger. This causes calls to the find image api to return [] and the calls are set up to run 10 times until a non empty result set is returned. These repeated calls to find image delay the dialog opening and the user is stuck watching a spinner for 10-20s. By switching to using c.imageDescription.repository, the find image calls return actual results and the dialog opens immediately.

If the dialog is coming from a clone server group action we should still use the ImageId. And if the dialog is from a deploy stage in a pipeline where the Image details come from a previous Find Image stage, then we just need to not run a query at all.