feat(kubernetes): Configure Service Account Name (#3864)
Allows configurable Service Account Name for Server Groups and Run Job
configuration to enable applications access to the Kubernetes API if
need be. Default behavior is to assign the `default` Service Account if
left blank. I also added it to the Server Group details. Note, if the
service account is `default`, it will not show up as the API doesn't
return it with the ReplicaSet/Deployment.